Refactor : extract suggestions code from Member entity to dedicated new Suggestion entity, extract tests in dedicated unit tests.
