+++ test/models/MemberTest.java
@@ -1,12 +1,8 @@
 package models;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
 
 import org.h2.util.StringUtils;
 import org.junit.*;
-import play.Logger;
 import play.test.*;
 
 /**
@@ -85,7 +81,7 @@ public class MemberTest extends UnitTest {
     }
 
     @Test
-    public void testInterests() {
+    public void testFindMembersInterestedIn() {
         Member bob = Member.findByLogin("bob");
         Member ced = Member.findByLogin("ced");
 
@@ -100,52 +96,8 @@ public class MemberTest extends UnitTest {
         assertEquals(2, Member.findMembersInterestedIn("Java").size());
         assertEquals(1, Member.findMembersInterestedIn("TDD").size());
         assertEquals(1, Member.findMembersInterestedIn("Hadoop").size());
-
-        // Advanced Checks
-        // Members inerested in ALL OF the Interests
-        List<Interest> interests_list1 = new ArrayList<Interest>();
-        interests_list1.add(Interest.findOrCreateByName("Java"));
-        assertEquals(2, Member.findMembersInterestedInAllOf(interests_list1).size());
-        interests_list1.add(Interest.findOrCreateByName("Hadoop"));
-        assertEquals(1, Member.findMembersInterestedInAllOf(interests_list1).size());
-        interests_list1.add(Interest.findOrCreateByName("TDD"));
-        assertEquals(0, Member.findMembersInterestedInAllOf(interests_list1).size());
-        
-        // Members inerested in AT LEAST ONE OF the Interests        
-        List<Interest> interests_list2 = new ArrayList<Interest>();
-        interests_list2.add(Interest.findOrCreateByName("Java"));
-        assertEquals(2, Member.findMembersInterestedInOneOf(interests_list2).size());
-        interests_list2.add(Interest.findOrCreateByName("Hadoop"));
-        assertEquals(2, Member.findMembersInterestedInOneOf(interests_list2).size());
-        interests_list2.add(Interest.findOrCreateByName("TDD"));
-        assertEquals(2, Member.findMembersInterestedInOneOf(interests_list2).size());
-
-        // Check Interests Cloud
-        // Be careful to the alphabetical order!
-        List<Map> cloud = Interest.getCloud();
-        assertEquals(
-                "[{interest=Hadoop, pound=1}, {interest=Java, pound=2}, {interest=TDD, pound=1}]",
-                cloud.toString());
-
     }
-    
-    @Test
-    public void testSuggestedMembers() {
-        Member bob = Member.findByLogin("bob");
-        Member ced = Member.findByLogin("ced");
-
-
-        // Add interest now : Java is common
-        ced.addInterest("Java").addInterest("Hadoop").save();
-        bob.addInterest("TDD").addInterest("Java").save();
-        
-
-        // Ced has already Bob in his links, so we except 0 suggested member
-        assertEquals(0, Member.suggestedMembersFor(ced).size());
-        // Bos hasn't already Ced in his links, so we except 1 suggested members
-        assertEquals(1, Member.suggestedMembersFor(bob).size());
-    }
-    
+   
     @Test
     public void testAddAccount() {
         Member bob = Member.findByLogin("bob");
